c bon

+++ src/raytracing.c
@@ -3,22 +3,20 @@
 /*                                                        :::      ::::::::   */
 /*   raytracing.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mjameau <mjameau@student.42.fr>            +#+  +:+       +#+        */
+/*   By: jchen <jchen@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/18 13:06:47 by mjameau           #+#    #+#             */
-/*   Updated: 2025/01/19 11:14:03 by mjameau          ###   ########.fr       */
+/*   Updated: 2025/01/19 13:18:21 by jchen            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../inc/cub3d.h"
 
-// PENSER A INIT TOUTES SES STRUCTURES (PLAYER ET RAY ET MACHINTRUC)
-
 // La fonction fabs prend un argument (un nombre a virgule, cf float)
 // et retourne sa valeur absolue en type double.
 static void	init_values(t_ray *ray, t_player *player)
 {
-	ray->cam_x = (2 * ray->curr_x) / WIN_WIDTH - 1;
+	ray->cam_x = (2 * ray->curr_x) / (double)WIN_WIDTH - 1;
 	ray->ray_dir_x = player->dir_vect.x + player->cam_plane_vect.x * ray->cam_x;
 	ray->ray_dir_y = player->dir_vect.y + player->cam_plane_vect.y * ray->cam_x;
 	ray->map_x = (int)player->px;
